Simplify code required from plugin
Push more stuff to the platform. Until the changes have made it into the SDK, they are implemented as a serious of nasty hacks.

ReSharper plugin tests now work too, as the default (test) .NET Framework platform provider, which provides platforms from the nuget packages, is patched to work on non-Windows.
